releasecab_api/requirements.txt
django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9 
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
